requirements.txt
playwright>=1.40.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
beautifulsoup4>=4.11.0
requests>=2.28.0
python-multipart>=0.0.6
transformers>=4.34.0